created test suite and moved delete functionality to its own file
